fix for unit jitter and basic unit grouping
unit Rigidbody is now set to 'Is Kinematic' to stop the Nav Mesh Agent
and Rigidbody fighting over unit movement. 'Priority' is now tracked so
units with fewer orders will move out of the way of units with more
orders.
Nav Mesh Agent's 'Stopping Distance' is set to be slightly smaller than
unitStoppingDistance.
A prefab for the test_unit has been created.